Create `VirtusizeUIBaseButton` for VirtusizeUIButton and VirtusizeUIRoundButton to inherit from
